refactor: test file path resolver
