+++ packages/reporter/src/utils/index.ts
@@ -1,1 +1,2 @@
 export { IssuesGrouper } from './issues-grouper';
+export { TestFilePathResolver } from './test-file-path-resolver';
