+++ packages/reporter/src/utils/test-file-path-resolver.ts
@@ -1,0 +1,26 @@
+import path from 'node:path';
+
+export class TestFilePathResolver {
+  public static getTestFilePath(): string {
+    // Check if running in Jest environment
+    const jestState = (global as any).expect?.getState?.();
+    if (jestState) {
+      const testPath = jestState.testPath;
+      const rootDir = jestState.snapshotState._rootDir;
+
+      return path.join(
+        path.basename(rootDir),
+        path.relative(rootDir, testPath)
+      );
+    }
+
+    // Relies on `TestContext` from Node.js built-in test runner appearing in the stack
+    const matchRes = String(new Error().stack).match(
+      /\n\s+at (?:async )?TestContext.* \((.*):\d+:\d+\)\n/
+    );
+
+    return matchRes?.[1]
+      ? path.relative(process.cwd(), matchRes[1] || '')
+      : 'unknown';
+  }
+}
